major refactor; seperating person and contact data interactions
