+++ internal/datastores/person_test.go
@@ -135,357 +135,6 @@ func Test_personStore_Add(t *testing.T) {
 	}
 }
 
-func Test_personStore_AddSpecificContactAddress(t *testing.T) {
-	type args struct {
-		ctx     context.Context
-		address models.ContactAddress
-	}
-	type wantQuery struct {
-		rawQuery  string
-		arguments []driver.Value
-		result    *sqlmock.Rows
-		returnErr error
-	}
-
-	testSqlBuilder := jagsqlb.NewSqlBuilder()
-	mockSession, mockDb, err := sqlmock.New()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	testPersonID := uuid.New()
-	newAddressID := uuid.New()
-	tests := []struct {
-		name      string
-		ps        personStore
-		args      args
-		wantQuery *wantQuery
-		wantID    uuid.UUID
-		assertion assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Success",
-			ps: personStore{
-				dbConn:       mockSession,
-				sqlBuilder:   testSqlBuilder,
-				tableNameMap: map[string]string{"addresses": "person.addresses"},
-			},
-			args: args{
-				ctx: context.Background(),
-				address: models.ContactAddress{
-					PersonID:   testPersonID,
-					Street1:    "123 Test Dr",
-					Street2:    "APT 1",
-					Locality:   "Testington",
-					Region:     "Testaria",
-					PostalCode: "12345-6789",
-					Country:    "Testopia",
-					Type:       models.AddressTypePhysical,
-					Primary:    true,
-				},
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `INSERT INTO "person"."addresses" ("person_id", "street1", "street2", "locality", "region", "postal_code", "country", "kind", "primary") VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9) RETURNING "id";`,
-				arguments: []driver.Value{testPersonID.String(), "123 Test Dr", "APT 1", "Testington", "Testaria", "12345-6789", "Testopia", "physical", true},
-				result:    sqlmock.NewRows([]string{"id"}).AddRow(newAddressID),
-			},
-			wantID:    newAddressID,
-			assertion: assert.NoError,
-		},
-		{
-			name: "Error; Query Builder",
-			ps: personStore{
-				dbConn:       mockSession,
-				sqlBuilder:   testSqlBuilder,
-				tableNameMap: map[string]string{"emails": ".bad_val"},
-			},
-			args: args{
-				ctx:     context.Background(),
-				address: models.ContactAddress{},
-			},
-			wantID:    uuid.Nil,
-			assertion: assert.Error,
-		},
-		{
-			name: "Error; Query Execution",
-			ps: personStore{
-				dbConn:       mockSession,
-				sqlBuilder:   testSqlBuilder,
-				tableNameMap: map[string]string{"addresses": "person.addresses"},
-			},
-			args: args{
-				ctx: context.Background(),
-				address: models.ContactAddress{
-					PersonID:   testPersonID,
-					Street1:    "123 Test Dr",
-					Street2:    "APT 1",
-					Locality:   "Testington",
-					Region:     "Testaria",
-					PostalCode: "12345-6789",
-					Country:    "Testopia",
-					Type:       models.AddressTypePhysical,
-					Primary:    true,
-				},
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `INSERT INTO "person"."addresses" ("person_id", "street1", "street2", "locality", "region", "postal_code", "country", "kind", "primary") VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9) RETURNING "id";`,
-				arguments: []driver.Value{testPersonID.String(), "123 Test Dr", "APT 1", "Testington", "Testaria", "12345-6789", "Testopia", "physical", true},
-				result:    sqlmock.NewRows(nil),
-				returnErr: assert.AnError,
-			},
-			assertion: assert.Error,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.wantQuery != nil {
-				mockDb.ExpectQuery(regexp.QuoteMeta(tt.wantQuery.rawQuery)).
-					WithArgs(tt.wantQuery.arguments...).WillReturnRows(
-					tt.wantQuery.result,
-				).WillReturnError(tt.wantQuery.returnErr)
-			}
-
-			got, err := tt.ps.AddSpecificContactAddress(tt.args.ctx, tt.args.address)
-			tt.assertion(t, err)
-			assert.Equal(t, tt.wantID, got)
-
-			// Check to see if the expectations of the query were met
-			if err := mockDb.ExpectationsWereMet(); err != nil {
-				t.Errorf("sql expectations were not met: %v", err)
-			}
-		})
-	}
-}
-
-func Test_personStore_AddSpecificContactEmail(t *testing.T) {
-	type args struct {
-		ctx   context.Context
-		email models.ContactEmail
-	}
-	type wantQuery struct {
-		rawQuery  string
-		arguments []driver.Value
-		result    *sqlmock.Rows
-		returnErr error
-	}
-
-	testSqlBuilder := jagsqlb.NewSqlBuilder()
-	mockSession, mockDb, err := sqlmock.New()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	testPersonID := uuid.New()
-	newEmailID := uuid.New()
-
-	tests := []struct {
-		name      string
-		ps        personStore
-		args      args
-		wantQuery *wantQuery
-		wantID    uuid.UUID
-		assertion assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Success",
-			ps: personStore{
-				dbConn:       mockSession,
-				sqlBuilder:   testSqlBuilder,
-				tableNameMap: map[string]string{"emails": "person.emails"},
-			},
-			args: args{
-				ctx: context.Background(),
-				email: models.ContactEmail{
-					PersonID: testPersonID,
-					Username: "test",
-					Provider: "example.com",
-					Primary:  true,
-				},
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `INSERT INTO "person"."emails" ("person_id", "username", "provider", "primary") VALUES ($1, $2, $3, $4) RETURNING "id";`,
-				arguments: []driver.Value{testPersonID.String(), "test", "example.com", true},
-				result:    sqlmock.NewRows([]string{"id"}).AddRow(newEmailID),
-			},
-			wantID:    newEmailID,
-			assertion: assert.NoError,
-		},
-		{
-			name: "Error; Query Builder",
-			ps: personStore{
-				dbConn:       mockSession,
-				sqlBuilder:   testSqlBuilder,
-				tableNameMap: map[string]string{"emails": ".bad_val"},
-			},
-			args: args{
-				ctx:   context.Background(),
-				email: models.ContactEmail{},
-			},
-			wantID:    uuid.Nil,
-			assertion: assert.Error,
-		},
-		{
-			name: "Error; Query Execution",
-			ps: personStore{
-				dbConn:       mockSession,
-				sqlBuilder:   testSqlBuilder,
-				tableNameMap: map[string]string{"emails": "person.emails"},
-			},
-			args: args{
-				ctx: context.Background(),
-				email: models.ContactEmail{
-					PersonID: testPersonID,
-					Username: "test",
-					Provider: "example.com",
-					Primary:  true,
-				},
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `INSERT INTO "person"."emails" ("person_id", "username", "provider", "primary") VALUES ($1, $2, $3, $4) RETURNING "id";`,
-				arguments: []driver.Value{testPersonID.String(), "test", "example.com", true},
-				result:    sqlmock.NewRows(nil),
-				returnErr: assert.AnError,
-			},
-			wantID:    uuid.Nil,
-			assertion: assert.Error,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.wantQuery != nil {
-				mockDb.ExpectQuery(regexp.QuoteMeta(tt.wantQuery.rawQuery)).
-					WithArgs(tt.wantQuery.arguments...).WillReturnRows(
-					tt.wantQuery.result,
-				).WillReturnError(tt.wantQuery.returnErr)
-			}
-
-			got, err := tt.ps.AddSpecificContactEmail(tt.args.ctx, tt.args.email)
-			tt.assertion(t, err)
-			assert.Equal(t, tt.wantID, got)
-
-			// Check to see if the expectations of the query were met
-			if err := mockDb.ExpectationsWereMet(); err != nil {
-				t.Errorf("sql expectations were not met: %v", err)
-			}
-		})
-	}
-}
-
-func Test_personStore_AddSpecificContactPhone(t *testing.T) {
-	type args struct {
-		ctx   context.Context
-		phone models.ContactPhone
-	}
-	type wantQuery struct {
-		rawQuery  string
-		arguments []driver.Value
-		result    *sqlmock.Rows
-		returnErr error
-	}
-
-	testSqlBuilder := jagsqlb.NewSqlBuilder()
-	mockSession, mockDb, err := sqlmock.New()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	testPersonID := uuid.New()
-	newPhoneID := uuid.New()
-
-	tests := []struct {
-		name      string
-		ps        personStore
-		args      args
-		wantQuery *wantQuery
-		wantID    uuid.UUID
-		assertion assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Success",
-			ps: personStore{
-				dbConn:       mockSession,
-				sqlBuilder:   testSqlBuilder,
-				tableNameMap: map[string]string{"phones": "person.phones"},
-			},
-			args: args{
-				ctx: context.Background(),
-				phone: models.ContactPhone{
-					PersonID:    testPersonID,
-					CountryCode: 1,
-					PhoneNumber: "(555) 555-5555",
-					Type:        models.PhoneTypeHome,
-					Primary:     true,
-				},
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `INSERT INTO "person"."phones" ("person_id", "country_code", "phone_number", "kind", "primary") VALUES ($1, $2, $3, $4, $5) RETURNING "id";`,
-				arguments: []driver.Value{testPersonID.String(), 1, "(555) 555-5555", "home", true},
-				result:    sqlmock.NewRows([]string{"id"}).AddRow(newPhoneID),
-			},
-			wantID:    newPhoneID,
-			assertion: assert.NoError,
-		},
-		{
-			name: "Error; Query Builder",
-			ps: personStore{
-				dbConn:       mockSession,
-				sqlBuilder:   testSqlBuilder,
-				tableNameMap: map[string]string{"emails": ".bad_val"},
-			},
-			args: args{
-				ctx:   context.Background(),
-				phone: models.ContactPhone{},
-			},
-			wantID:    uuid.Nil,
-			assertion: assert.Error,
-		},
-		{
-			name: "Error; Query Execution",
-			ps: personStore{
-				dbConn:       mockSession,
-				sqlBuilder:   testSqlBuilder,
-				tableNameMap: map[string]string{"phones": "person.phones"},
-			},
-			args: args{
-				ctx: context.Background(),
-				phone: models.ContactPhone{
-					PersonID:    testPersonID,
-					CountryCode: 1,
-					PhoneNumber: "(555) 555-5555",
-					Type:        models.PhoneTypeHome,
-					Primary:     true,
-				},
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `INSERT INTO "person"."phones" ("person_id", "country_code", "phone_number", "kind", "primary") VALUES ($1, $2, $3, $4, $5) RETURNING "id";`,
-				arguments: []driver.Value{testPersonID.String(), 1, "(555) 555-5555", "home", true},
-				result:    sqlmock.NewRows(nil),
-				returnErr: assert.AnError,
-			},
-			assertion: assert.Error,
-		},
-	}
-	for _, tt := range tests {
-		if tt.wantQuery != nil {
-			mockDb.ExpectQuery(regexp.QuoteMeta(tt.wantQuery.rawQuery)).
-				WithArgs(tt.wantQuery.arguments...).WillReturnRows(
-				tt.wantQuery.result,
-			).WillReturnError(tt.wantQuery.returnErr)
-		}
-
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := tt.ps.AddSpecificContactPhone(tt.args.ctx, tt.args.phone)
-			tt.assertion(t, err)
-			assert.Equal(t, tt.wantID, got)
-		})
-
-		// Check to see if the expectations of the query were met
-		if err := mockDb.ExpectationsWereMet(); err != nil {
-			t.Errorf("sql expectations were not met: %v", err)
-		}
-	}
-}
-
 func Test_personStore_Delete(t *testing.T) {
 	type args struct {
 		ctx context.Context
@@ -1082,404 +731,6 @@ func Test_personStore_Update(t *testing.T) {
 			}
 
 			tt.assertion(t, tt.ps.Update(tt.args.ctx, tt.args.id, tt.args.item))
-
-			// Check to see if the expectations of the query were met
-			if err := mockDb.ExpectationsWereMet(); err != nil {
-				t.Errorf("sql expectations were not met: %v", err)
-			}
-		})
-	}
-}
-
-func Test_personStore_GetSpecificContactAddresses(t *testing.T) {
-	type args struct {
-		ctx context.Context
-		id  uuid.UUID
-	}
-
-	type wantQuery struct {
-		rawQuery  string
-		arguments []driver.Value
-		result    *sqlmock.Rows
-		returnErr error
-	}
-
-	addressColumns := []string{"id", "person_id", "street1", "street2", "locality", "region", "postal_code", "country", "kind", "primary"}
-	testSqlBuilder := jagsqlb.NewSqlBuilder()
-	mockSession, mockDb, err := sqlmock.New()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	testPersonStore := personStore{
-		dbConn:     mockSession,
-		sqlBuilder: testSqlBuilder,
-		tableNameMap: map[string]string{
-			"addresses": "person.addresses",
-		},
-	}
-	testPersonID := uuid.New()
-	testNotFoundID := uuid.New()
-	testAddressID1 := uuid.New()
-	testAddressID2 := uuid.New()
-
-	tests := []struct {
-		name      string
-		ps        personStore
-		args      args
-		wantQuery *wantQuery
-		want      []models.ContactAddress
-		assertion assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Success",
-			ps:   testPersonStore,
-			args: args{
-				ctx: context.Background(),
-				id:  testPersonID,
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `SELECT * FROM "person"."addresses" WHERE "person_id" = $1`,
-				arguments: []driver.Value{testPersonID.String()},
-				result: sqlmock.NewRows(addressColumns).AddRows(
-					[]driver.Value{
-						testAddressID1, testPersonID, "123 Test Dr", "APT 1", "Testerville", "Testton", "12345-6789", "U.S.", "physical", true,
-					},
-					[]driver.Value{
-						testAddressID2, testPersonID, "P.O. Box 123", "", "Testerville", "Testton", "12344-5567", "U.S.", "mailing", true,
-					},
-				),
-			},
-			want: []models.ContactAddress{
-				{
-					ID:         testAddressID1,
-					PersonID:   testPersonID,
-					Street1:    "123 Test Dr",
-					Street2:    "APT 1",
-					Locality:   "Testerville",
-					Region:     "Testton",
-					PostalCode: "12345-6789",
-					Country:    "U.S.",
-					Type:       "physical",
-					Primary:    true,
-				},
-				{
-					ID:         testAddressID2,
-					PersonID:   testPersonID,
-					Street1:    "P.O. Box 123",
-					Locality:   "Testerville",
-					Region:     "Testton",
-					PostalCode: "12344-5567",
-					Country:    "U.S.",
-					Type:       "mailing",
-					Primary:    true,
-				},
-			},
-			assertion: assert.NoError,
-		},
-		{
-			name: "Error; SQL Builder",
-			ps: personStore{
-				dbConn:     mockSession,
-				sqlBuilder: testSqlBuilder,
-				tableNameMap: map[string]string{
-					"emails": ".bad_val",
-				},
-			},
-			args: args{
-				ctx: context.Background(),
-				id:  testPersonID,
-			},
-			assertion: assert.Error,
-		},
-		{
-			name: "Error; SQL Execution",
-			ps:   testPersonStore,
-			args: args{
-				ctx: context.Background(),
-				id:  testNotFoundID,
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `SELECT * FROM "person"."addresses" WHERE "person_id" = $1`,
-				arguments: []driver.Value{testNotFoundID.String()},
-				result:    sqlmock.NewRows(nil),
-				returnErr: assert.AnError,
-			},
-			assertion: assert.Error,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.wantQuery != nil {
-				mockDb.ExpectQuery(regexp.QuoteMeta(tt.wantQuery.rawQuery)).
-					WithArgs(tt.wantQuery.arguments...).WillReturnRows(
-					tt.wantQuery.result,
-				).WillReturnError(tt.wantQuery.returnErr)
-			}
-
-			got, err := tt.ps.GetSpecificContactAddresses(tt.args.ctx, tt.args.id)
-			tt.assertion(t, err)
-			assert.Equal(t, tt.want, got)
-
-			// Check to see if the expectations of the query were met
-			if err := mockDb.ExpectationsWereMet(); err != nil {
-				t.Errorf("sql expectations were not met: %v", err)
-			}
-		})
-	}
-}
-
-func Test_personStore_GetSpecificContactEmails(t *testing.T) {
-	type args struct {
-		ctx context.Context
-		id  uuid.UUID
-	}
-
-	type wantQuery struct {
-		rawQuery  string
-		arguments []driver.Value
-		result    *sqlmock.Rows
-		returnErr error
-	}
-
-	emailColumns := []string{"id", "person_id", "username", "provider", "primary"}
-	testSqlBuilder := jagsqlb.NewSqlBuilder()
-	mockSession, mockDb, err := sqlmock.New()
-	if err != nil {
-		t.Fatal(err)
-	}
-	testPersonID := uuid.New()
-	testNotFoundID := uuid.New()
-	testEmailID1 := uuid.New()
-	testEmailID2 := uuid.New()
-
-	testPersonStore := personStore{
-		dbConn:     mockSession,
-		sqlBuilder: testSqlBuilder,
-		tableNameMap: map[string]string{
-			"emails": "person.emails",
-		},
-	}
-
-	tests := []struct {
-		name      string
-		ps        personStore
-		args      args
-		wantQuery *wantQuery
-		want      []models.ContactEmail
-		assertion assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Success",
-			ps:   testPersonStore,
-			args: args{
-				ctx: context.Background(),
-				id:  testPersonID,
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `SELECT * FROM "person"."emails" WHERE "person_id" = $1`,
-				arguments: []driver.Value{testPersonID.String()},
-				result: sqlmock.NewRows(emailColumns).AddRows(
-					[]driver.Value{
-						testEmailID1.String(), testPersonID.String(), "tester", "example.com", true,
-					},
-					[]driver.Value{
-						testEmailID2.String(), testPersonID.String(), "admin", "example.com", false,
-					},
-				),
-			},
-			want: []models.ContactEmail{
-				{
-					ID:       testEmailID1,
-					PersonID: testPersonID,
-					Username: "tester",
-					Provider: "example.com",
-					Primary:  true,
-				},
-				{
-					ID:       testEmailID2,
-					PersonID: testPersonID,
-					Username: "admin",
-					Provider: "example.com",
-					Primary:  false,
-				},
-			},
-			assertion: assert.NoError,
-		},
-		{
-			name: "Error; Query Builder",
-			ps: personStore{
-				dbConn:     mockSession,
-				sqlBuilder: testSqlBuilder,
-				tableNameMap: map[string]string{
-					"emails": ".bad_val",
-				},
-			},
-			args: args{
-				ctx: context.Background(),
-				id:  testPersonID,
-			},
-			assertion: assert.Error,
-		},
-		{
-			name: "Error; Query Execution",
-			ps:   testPersonStore,
-			args: args{
-				ctx: context.Background(),
-				id:  testNotFoundID,
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `SELECT * FROM "person"."emails" WHERE "person_id" = $1`,
-				arguments: []driver.Value{testNotFoundID.String()},
-				result:    sqlmock.NewRows(nil),
-				returnErr: assert.AnError,
-			},
-			assertion: assert.Error,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.wantQuery != nil {
-				mockDb.ExpectQuery(regexp.QuoteMeta(tt.wantQuery.rawQuery)).
-					WithArgs(tt.wantQuery.arguments...).WillReturnRows(
-					tt.wantQuery.result,
-				).WillReturnError(tt.wantQuery.returnErr)
-			}
-
-			got, err := tt.ps.GetSpecificContactEmails(tt.args.ctx, tt.args.id)
-			tt.assertion(t, err)
-			assert.Equal(t, tt.want, got)
-
-			// Check to see if the expectations of the query were met
-			if err := mockDb.ExpectationsWereMet(); err != nil {
-				t.Errorf("sql expectations were not met: %v", err)
-			}
-		})
-	}
-}
-
-func Test_personStore_GetSpecificContactPhones(t *testing.T) {
-	type args struct {
-		ctx context.Context
-		id  uuid.UUID
-	}
-
-	type wantQuery struct {
-		rawQuery  string
-		arguments []driver.Value
-		result    *sqlmock.Rows
-		returnErr error
-	}
-
-	phoneColumns := []string{"id", "person_id", "country_code", "phone_number", "kind", "primary"}
-	testSqlBuilder := jagsqlb.NewSqlBuilder()
-	mockSession, mockDb, err := sqlmock.New()
-	if err != nil {
-		t.Fatal(err)
-	}
-	testPersonID := uuid.New()
-	testNotFoundID := uuid.New()
-	testPhoneID1 := uuid.New()
-	testPhoneID2 := uuid.New()
-
-	testPersonStore := personStore{
-		dbConn:     mockSession,
-		sqlBuilder: testSqlBuilder,
-		tableNameMap: map[string]string{
-			"phones": "person.phones",
-		},
-	}
-
-	tests := []struct {
-		name      string
-		ps        personStore
-		args      args
-		wantQuery *wantQuery
-		want      []models.ContactPhone
-		assertion assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Success",
-			ps:   testPersonStore,
-			args: args{
-				ctx: context.Background(),
-				id:  testPersonID,
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `SELECT * FROM "person"."phones" WHERE "person_id" = $1`,
-				arguments: []driver.Value{testPersonID.String()},
-				result: sqlmock.NewRows(phoneColumns).AddRows(
-					[]driver.Value{
-						testPhoneID1.String(), testPersonID.String(), 1, "(315)555-1234", "home", true,
-					},
-					[]driver.Value{
-						testPhoneID2.String(), testPersonID.String(), 1, "(315)555-9876", "cell", false,
-					},
-				),
-			},
-			want: []models.ContactPhone{
-				{
-					ID:          testPhoneID1,
-					PersonID:    testPersonID,
-					CountryCode: 1,
-					PhoneNumber: "(315)555-1234",
-					Type:        "home",
-					Primary:     true,
-				},
-				{
-					ID:          testPhoneID2,
-					PersonID:    testPersonID,
-					CountryCode: 1,
-					PhoneNumber: "(315)555-9876",
-					Type:        "cell",
-					Primary:     false,
-				},
-			},
-			assertion: assert.NoError,
-		},
-		{
-			name: "Error; SQL Builder",
-			ps: personStore{
-				dbConn:     mockSession,
-				sqlBuilder: testSqlBuilder,
-				tableNameMap: map[string]string{
-					"phone": ".bad_val",
-				},
-			},
-			args: args{
-				ctx: context.Background(),
-				id:  testPersonID,
-			},
-			assertion: assert.Error,
-		},
-		{
-			name: "Error; SQL Execution",
-			ps:   testPersonStore,
-			args: args{
-				ctx: context.Background(),
-				id:  testNotFoundID,
-			},
-			wantQuery: &wantQuery{
-				rawQuery:  `SELECT * FROM "person"."phones" WHERE "person_id" = $1`,
-				arguments: []driver.Value{testNotFoundID.String()},
-				result:    sqlmock.NewRows(nil),
-				returnErr: assert.AnError,
-			},
-			assertion: assert.Error,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.wantQuery != nil {
-				mockDb.ExpectQuery(regexp.QuoteMeta(tt.wantQuery.rawQuery)).
-					WithArgs(tt.wantQuery.arguments...).WillReturnRows(
-					tt.wantQuery.result,
-				).WillReturnError(tt.wantQuery.returnErr)
-			}
-
-			got, err := tt.ps.GetSpecificContactPhones(tt.args.ctx, tt.args.id)
-			tt.assertion(t, err)
-			assert.Equal(t, tt.want, got)
 
 			// Check to see if the expectations of the query were met
 			if err := mockDb.ExpectationsWereMet(); err != nil {
